+++ editors/neovim/ftplugin/goop.lua
@@ -1,5 +1,5 @@
 vim.bo.commentstring = "// %s"
-vim.bo.comments = "s1:(*,mb:*,ex:*),://"
+vim.bo.comments = "://"
 vim.bo.expandtab = true
 vim.bo.shiftwidth = 2
 vim.bo.tabstop = 2